Change f64 to f32

+++ src/v1_6/types/charging_schedule.rs
@@ -18,5 +18,5 @@ pub struct ChargingSchedule {
     pub charging_schedule_period: Vec<ChargingSchedulePeriod>,
     /// Optional. Minimum charging rate supported by the electric vehicle. The unit of measure is defined by the chargingRateUnit. This parameter is intended to be used by a local smart charging algorithm to optimize the power allocation for in the case a charging process is inefficient at lower charging rates. Accepts at most one digit fraction (e.g. 8.1)
     #[serde(skip_serializing_if = "Option::is_none")]
-    pub min_charging_rate: Option<f64>,
+    pub min_charging_rate: Option<f32>,
 }
