+++ src/v2_0_1/datatypes/charging_schedule_period_type.rs
@@ -6,7 +6,7 @@ pub struct ChargingSchedulePeriodType {
     /// Required. Start of the period, in seconds from the start of schedule. The value of StartPeriod also defines the stop time of the previous period.
     pub start_period: i32,
     /// Required. Charging rate limit during the schedule period, in the applicable chargingRateUnit, for example in Amperes (A) or Watts (W). Accepts at most one digit fraction (e.g. 8.1).
-    pub limit: f64,
+    pub limit: f32,
     /// Optional. The number of phases that can be used for charging. If a number of phases is needed, numberPhases=3 will be assumed unless another number is given.
     #[serde(skip_serializing_if = "Option::is_none")]
     pub number_phases: Option<i32>,
